[artf6549]: Deleting repos from CTF mode does not delete the repo but only archives it.
* integration/ctf/webapp/src/main/java/com/vasoftware/sf/common/util/FileUtil.java
  (deleteDir): If initial delete does not work,  a gc and short wait can clear it up,
  if it results from .nfsXXX files.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2554 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ ctf/webapp/src/main/java/com/vasoftware/sf/common/util/FileUtil.java
@@ -78,7 +78,24 @@ public class FileUtil {
             }
         }
 
-        return dir.delete();
+        boolean isDeleted = dir.delete();
+        if (!isDeleted && dir.exists()) {
+            // Directories mounted on NFS volumes may have lingering .nfsXXXX files 
+            // pointing to deleted files which are still referenced by the JVM.  We don't
+            // explicitly have any handles open, however it appears some are created during
+            // the listing above, or in some other unknown way. It seems it is possible to
+            // free them by cleaning up stale objects, and giving the OS time to cleanup as 
+            // well.
+            System.gc();
+            try {
+                Thread.sleep(100);                
+            } catch (InterruptedException e) {
+                // ignored
+            }
+            isDeleted = dir.delete();
+        }
+        
+        return isDeleted;
     }
 
     /**
